# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API для аренды квартир",
        "title": "Renting App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8000/api",
    "basePath": "/",
    "paths": {
        "/advert": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Создает объявление авторизированному пользователю",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Создать объявление",
                "parameters": [
                    {
                        "description": "Advert input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateAdvertInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created"
                    },
                    "400": {
                        "description": "Некорректный ID объявления",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    }
                }
            }
        },
        "/advert/{id}": {
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Удалить объявление",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Advert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Обновить объявление",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Advert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update input",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateAdvertInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/adverts": {
            "get": {
                "description": "Возвращает список всех объявлений с фильтрацией. Если пользователь авторизован — добавляет флаг избранного.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Получение списка объявлений",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Город",
                        "name": "city",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Район",
                        "name": "district",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Тип санузла",
                        "name": "bathroom_type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Ремонт",
                        "name": "remont",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Тип аренды",
                        "name": "rental_type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Количество комнат",
                        "name": "rooms",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Минимальная цена",
                        "name": "price_from",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Максимальная цена",
                        "name": "price_to",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Минимальный этаж",
                        "name": "floor_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Максимальный этаж",
                        "name": "floor_max",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Минимальный год постройки",
                        "name": "year_min",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Максимальный год постройки",
                        "name": "year_max",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Минимальный рейтинг квартиры",
                        "name": "apartment_rating_min",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Минимальный рейтинг арендодателя",
                        "name": "landlord_rating_min",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие лифта",
                        "name": "elevator",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие консьержа",
                        "name": "concierge",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Можно с животными",
                        "name": "pets",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Можно с детьми",
                        "name": "babies",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Можно курить",
                        "name": "smoking",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие интернета",
                        "name": "internet",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие стиральной машины",
                        "name": "washing_machine",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие телевизора",
                        "name": "tv",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие кондиционера",
                        "name": "conditioner",
                        "in": "query"
                    },
                    {
                        "type": "boolean",
                        "description": "Наличие посудомойки",
                        "name": "dishwasher",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Широта для геолокации",
                        "name": "lat",
                        "in": "query"
                    },
                    {
                        "type": "number",
                        "description": "Долгота для геолокации",
                        "name": "lng",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Лимит записей, по умолчанию 20",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Смещение, по умолчанию 0",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Поле сортировки, по умолчанию created_at",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Направление сортировки: asc или desc, по умолчанию desc",
                        "name": "order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.AdvertListResponse"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    }
                }
            }
        },
        "/adverts/{id}": {
            "get": {
                "description": "Возвращает подробности одного объявления по его идентификатору",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Получение объявления по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID объявления",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAdvertResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный ID объявления",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    }
                }
            }
        },
...
        "/my/advert/{id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Возвращает объявление пользователя по ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "adverts"
                ],
                "summary": "Получить объявление пользователя по ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Advert ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.GetAdvertResponse"
                        }
                    },
                    "400": {
                        "description": "Некорректный ID объявления",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    },
                    "401": {
                        "description": "Пользователь не авторизован",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка сервера",
                        "schema": {
                            "$ref": "#/definitions/error.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateAdvertInput": {
            "type": "object",
            "required": [
                "apartment_id",
                "deposit",
                "rent",
                "rental_type",
                "title"
            ],
            "properties": {
                "apartment_id": {
                    "type": "integer"
                },
                "babies": {
                    "type": "boolean"
                },
                "concierge": {
                    "type": "boolean"
                },
                "conditioner": {
                    "type": "boolean"
                },
                "deposit": {
                    "type": "number"
                },
                "dishwasher": {
                    "type": "boolean"
                },
                "internet": {
                    "type": "boolean"
                },
                "pets": {
                    "type": "boolean"
                },
                "rent": {
                    "type": "number"
                },
                "rental_type": {
                    "type": "string"
                },
                "smoking": {
                    "type": "boolean"
                },
                "title": {
                    "type": "string"
                },
                "tv": {
                    "type": "boolean"
                },
                "washing_machine": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.GetAdvertResponse": {
            "type": "object",
            "properties": {
                "apartment": {
                    "$ref": "#/definitions/dto.GetApartmentResponse"
                },
                "apartment_id": {
                    "type": "integer"
                },
                "babies": {
                    "type": "boolean"
                },
                "concierge": {
                    "type": "boolean"
                },
                "conditioner": {
                    "type": "boolean"
                },
                "created_at": {
                    "type": "string"
                },
                "deposit": {
                    "type": "number"
                },
                "dishwasher": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "internet": {
                    "type": "boolean"
                },
                "is_favorite": {
                    "type": "boolean"
                },
                "pets": {
                    "type": "boolean"
                },
                "rent": {
                    "type": "number"
                },
                "rental_type": {
                    "type": "string"
                },
                "smoking": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tv": {
                    "type": "boolean"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                },
                "washing_machine": {
                    "type": "boolean"
                }
            }
        },
...
        "dto.UpdateAdvertInput": {
            "type": "object",
            "properties": {
                "babies": {
                    "type": "boolean"
                },
                "concierge": {
                    "type": "boolean"
                },
                "conditioner": {
                    "type": "boolean"
                },
                "deposit": {
                    "type": "number"
                },
                "dishwasher": {
                    "type": "boolean"
                },
                "internet": {
                    "type": "boolean"
                },
                "pets": {
                    "type": "boolean"
                },
                "rent": {
                    "type": "number"
                },
                "rental_type": {
                    "type": "string"
                },
                "smoking": {
                    "type": "boolean"
                },
                "status": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "tv": {
                    "type": "boolean"
                },
                "washing_machine": {
                    "type": "boolean"
                }
            }
        },
...
        "error.Response": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/error.ErrorDetail"
                    }
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "handler.AdvertListResponse": {
            "type": "object",
            "properties": {
                "adverts": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/dto.GetAdvertResponse"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
        "handler.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Invalid input data"
                }
            }
        },
...
    }
}